+++ Sources/TextMessageCell.swift
@@ -70,6 +70,9 @@ open class TextMessageCell: MessageCollectionViewCell<MessageLabel> {
             messageContentView.text = text
         case .attributedText(let text):
             messageContentView.attributedText = text
+        case .emoji(let text):
+            messageContentView.text = text
+            messageContentView.font = messageContentView.font.withSize(2 * messageContentView.font.pointSize)
         default:
             break
         }